refactor(core): split config_models by layer modules
